+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 5.3.4 (Next)
 
-* Your contribution here.
+* [#241](https://github.com/mongoid/mongoid-slug/pull/240): Fix: ensure localized slugs index expected localized slug fields - [@onomated](https://github.com/onomated)
 
 ## 5.3.3 (2017/04/06)